Finished with the rough draft of the new initial guess function for str_root() and switching over to str_root(string, double) in lieu of str_root(string, string) -- did the same for str_pow() too. Got it to compile again, but still have some computing glitches.

+++ library/bgnm.hpp
@@ -245,7 +245,7 @@ class Bgnm
     void operator %= (const N & number);
     void operator %= (const char * cs);
     
-    Bgnm        root(const float & index) const;
+    Bgnm        root(const double & index) const;
     Bgnm        sqrt()  const;     // return square root
     Bgnm        cbrt()  const;     // return cube root
     Bgnm        abs()   const;     // return absolute value
